requirements.txt
flask==3.0.0
tensorflow==2.17.0
opencv-python-headless==4.8.1.78
numpy==1.26.4
pillow==10.2.0
werkzeug==3.0.1
